refactor: move recruitment page to app router
